refactor(core,schemas): move webhook event payload type definition to schemas
move webhook event payload type definition to schemas
